+++ docs/changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.6.1 (unreleased to Maven Central)
+
+- Fixes #15 (PlantUML export fails when element names are Unicode characters)
+
 ## 1.6.0 (15th August 2022)
 
 - `com.structurizr.export.DiagramExporter` and `com.structurizr.export.WorkspaceExporter` can now be implemented to build custom exporters, for use with the Structurizr CLI `export` command.